requirements.txt
robotframework

black
isort
flake8
pylint
abstract-singleton
wheel
setuptools
build
twine